+++ eyes.sdk.core/src/main/java/com/applitools/eyes/visualgrid/services/RunnerOptions.java
@@ -4,6 +4,11 @@ import com.applitools.eyes.AbstractProxySettings;
 
 public class RunnerOptions {
 
+    public enum AutProxyMode {
+        BLOCK,
+        ALLOW
+    }
+
     private Integer testConcurrency = null;
     private String apiKey = null;
     private String serverUrl = null;
@@ -12,6 +17,7 @@ public class RunnerOptions {
     private boolean isAutProxySet = false;
     private AbstractProxySettings autProxy = null;
     private String[] autProxyDomains = null;
+    private AutProxyMode autProxyMode;
 
     public RunnerOptions testConcurrency(int testConcurrency) {
         this.testConcurrency = testConcurrency;
@@ -61,14 +67,25 @@ public class RunnerOptions {
     }
 
     /**
-     * Setting a separated proxy for requests made to the given domains. Requests to other non-eyes domains will be sent without a proxy.
+     * Setting a separated proxy for requests sent to the given domains. Requests to other non-eyes domains will be sent without a proxy.
      * If the AUT proxy is set, it cannot be changed again for those specific requests by any means.
      * If AUT proxy is set to null, the behavior will be the same as {@link #autProxy(AbstractProxySettings)}
      */
     public RunnerOptions autProxy(AbstractProxySettings autProxy, String[] domains) {
+        return autProxy(autProxy, domains, AutProxyMode.ALLOW);
+    }
+
+    /**
+     * Setting a separated proxy for requests sent to the given domains (if mode is {@link AutProxyMode#ALLOW}
+     * or for requests sent to domains other than the given domains (if mode is {@link AutProxyMode#BLOCK}.
+     * If the AUT proxy is set, it cannot be changed again for those specific requests by any means.
+     * If AUT proxy is set to null, the behavior will be the same as {@link #autProxy(AbstractProxySettings)}
+     */
+    public RunnerOptions autProxy(AbstractProxySettings autProxy, String[] domains, AutProxyMode mode) {
         isAutProxySet = true;
         this.autProxy = autProxy;
         this.autProxyDomains = domains;
+        this.autProxyMode = mode;
         return this;
     }
 
@@ -78,6 +95,10 @@ public class RunnerOptions {
 
     public String[] getAutProxyDomains() {
         return autProxyDomains;
+    }
+
+    public AutProxyMode getAutProxyMode() {
+        return autProxyMode;
     }
 
     public boolean isAutProxySet() {