+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 ### Updated 
 - Checking driver context to define if mobile device used while getting pixel ratio of the device. [Trello 2650](https://trello.com/c/elq5kosS)
 - Cleanup stitched image if real size is smaller than expected. [Trello 2552](https://trello.com/c/cxcv9edh)
+- It's now possible to disable autProxy for specific domains through the `RunnerOptions` class. [Trello 2604](https://trello.com/c/mNTo6koP)
 ### Fixed
 - If we fail downloading a resource, we send and empty resource with the correct status code.
 - Simple regions' location in tests with css stitching is now correct. [Trello 2655](https://trello.com/c/hTwRpY39)